requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
openai>=1.6.1,<2.0.0
chromadb==0.4.18
sentence-transformers==2.2.2
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2
yfinance==0.2.28
newsapi-python==0.2.6
feedparser==6.0.10
schedule==1.2.0
aiohttp==3.9.1
websockets==12.0
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1 